streamlit/requirements.txt
altair==5.2.0
APScheduler==3.10.4
mariadb==1.1.10
matplotlib==3.8.3
openpyxl==3.1.2
pandas==2.2.0
pysftp==0.2.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
SQLAlchemy==2.0.27
st-pages==0.4.5
streamlit==1.32.1
tornado==6.4